feat: blocks have titles
